### 2024-08-27 #87 @Mathilde-Reynes
[Re] Model of thalamocortical slow-wave sleep oscillations and transitions to activated states
**Original article:** https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6757797/

**Scientific domain:**  Computational Neuroscience
**Programming language:**  Python (brian2)
**Suggested editor:**  [Benoît Girard](https://github.com/benoit-girard)
**Suggested reviewers:**  [Qihong Lu](https://github.com/qihongl); [Ozan Caglayan](https://github.com/ozancaglayan); [Pierre Enel](https://github.com/p-enel)
