### 2024-10-11 #87 @rougier
@benoit-girard Can you edit this submission, I think it's a perfect fit for you.
