### 2024-11-20 #87 @rougier
@benoit-girard  @apdavison @heplesser @vitay @mstimberg @pietromarchesi @stephanmg

Can your review this submission on "Model of thalamocortical slow-wave sleep oscillations and transitions to activated states"
