### 2024-12-19 #87 @apdavison
@rougier I'm afraid I don't think it will be possible before Christmas. I've pencilled this in for the first full week in January, I'm targeting January 10th.
