### 2024-12-19 #87 @stephanmg
> @benoit-girard @apdavison @heplesser @vitay @mstimberg @pietromarchesi @stephanmg
>
> Can your review this submission on "Model of thalamocortical slow-wave sleep oscillations and transitions to activated states"

terribly sorry, somehow this went under the radar for me, if help is still needed, happy to assist!
