### 2025-01-10 #87 @apdavison
I have started reviewing the submission, and I will post more complete comments soon; however, I would like here to make a preliminary suggestion which will both (i) help me finish the review and (ii) make the re-implemented model more reproducible.

I was able to run `python Thalamo_cortical.py`, which completes without error. The problem is that to reproduce the data for the figures, I need to uncomment certain lines in `Thalamo_cortical.py` and, for some figures, also in other files. It takes a lot of reading and checking code to make sure the correct lines have been uncommented and re-commented.

If the authors have time, it would be very helpful to put the commented-out lines into a dictionary, or into functions, so that all the changes are controlled by a single parameter (ideally provided on the command-line, e.g., `python Thalamo_cortical.py Figure4`). Similarly, saving the raw data should be controlled by another parameter.
