### 2025-02-06 028d846
function created to config params, save data & plot
