+++ Model/Figure_plotting.py
@@ -1,0 +1,654 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+from brian2 import *
+import numpy as np
+import matplotlib.pyplot as plt
+from brian2.units.constants import *
+import matplotlib.gridspec as gridspec
+from scipy import signal
+
+def figure_plotting(fig_number, all_monitors, all_monitors_T, all_synapses, monitor_poisson, runtime, N):
+    V1_PYd, V2_PYs, V3_INd, V4_INs, R2_PYs, R4_INs, I1_PYd, I2_INd, S1, S2, M0, M1 = all_monitors
+    V1_RE, V2_TC, R1_RE, R2_TC, I1_RE, I2_TC = all_monitors_T
+    S_AMPA_PY_PY, S_AMPA_PY_IN, S_NMDA_PY_PY, S_NMDA_PY_IN, S_GABAA_IN_PY = all_synapses
+    
+    # Figure 5 
+    if fig_number == "5" :
+        fig = plt.figure(figsize=(20, 12))
+        gs = gridspec.GridSpec(2, 2, width_ratios=[20, 0.5], height_ratios=[2, 1], wspace=0.05, hspace=0.3)
+        ax1 = plt.subplot(gs[0, 0])
+        im1 = ax1.imshow(V2_PYs.v/mV, aspect='auto', cmap='YlGnBu', vmax=-60, vmin=-75, 
+                     extent=[0, runtime, N-1, 0], interpolation='bicubic')
+        ax1.set_title('PY', size=30, loc='left')
+        ax1.set_ylabel('Neuron index', size=30, labelpad=30)
+        ax1.set_xlim(0, 30)
+        ax1.yaxis.set_major_locator(MultipleLocator(base=25))
+        ax1.tick_params(axis='both', which='major', labelsize=25, width=2)
+        ax2 = plt.subplot(gs[1, 0])
+        im2 = ax2.imshow(V2_TC.v/mV, aspect='auto', cmap='YlGnBu', vmax=-60, vmin=-75, 
+                    extent=[0, runtime, 50, 0], interpolation='bicubic')
+        ax2.set_title('TC', size=30, loc='left')
+        ax2.set_xlabel('Time (s)', size=30, labelpad=10)
+        ax2.set_ylabel('Neuron index', size=30, labelpad=30)
+        ax2.set_xlim(0, 30)
+        ax2.yaxis.set_major_locator(MultipleLocator(base=25))
+        ax2.tick_params(axis='both', which='major', labelsize=25, width=2)
+    
+        cbar_ax = plt.subplot(gs[:, 1])
+        cbar = fig.colorbar(im1, cax=cbar_ax, orientation='vertical')
+        cbar.set_label('Membrane potential (mV)', size=30, labelpad=30)
+        cbar.ax.tick_params(labelsize=25, width=2)
+        plt.show()
+        
+    # Figure 7 
+    if fig_number == "7" :
+        fig,ax = subplots(4,1, sharex = True,figsize=(19,18))
+        ax[0].plot(V2_PYs.t, V2_PYs.v[N//2]/mV,color="tab:blue",linewidth=1.5)
+        ax[0].set_title('PY', size=30, loc='left')
+        ax[0].set_ylabel('mV',size=30, labelpad=30)
+        ax[0].set_xlim([15,30])
+        ax[0].spines['top'].set_visible(False)
+        ax[0].spines['right'].set_visible(False)
+        ax[0].spines['bottom'].set_visible(True)
+        ax[0].spines['left'].set_visible(True)
+        ax[0].yaxis.set_major_locator(MultipleLocator(base=25))
[…]
