### 2025-02-06 #87 @Mathilde-Reynes
Dear all,

The changes proposed have been implemented. The git is up to date. In the meantime, we have identified an error in Figure 10, for which the middle and bottom panels were swapped (it was just a mistake in their position as the legend didn't match the figure). This issue has been corrected, and the revised manuscript has been updated as well.

Best
