### 2025-02-17 #87 @Mathilde-Reynes
Hello, please find below the changes implemented:

1. One function controls everything necessary to run the model, found within Thalamo_cortical.py; `thalamocortical_network(seed_value,analyze_speed,fig_number,raw_data,plot_figure) `with:
- analyze_speed (True/False) allows the user to compute the mean and std propagation speed of cortical up-states
- fig_number (str such as "9-A1","5" etc.) allows the user to run the simulation with the parameters necessary to reproduce a specific figure from the article
- raw_data (True/False) allows the user to select if raw data (.txt files) should be saved
- plot_figure (True/False) allows the user to plot the figure corresponding to the fig_number

2. Addition of .py files containing functions necessary for 1. :
- Figure_conditions.py : Designed to configure the model with specific parameters adjusted for different figures.
- Figure_plotting.py : Designed to plot a specific figure of the article.
- Figure_rawdata.py : Designed to save the raw data necessary for a specific figure.

3. py files to plot every single figure separately from the raw data were removed as everything is now integrated in Figure_plotting.py. Figure 2, 3, 4 remain as they are based off raw data from the original model.

4. Figure 10 was updated in a new file called [Re]ReynesAussel2024_erratum.pdf.  As previously said, the middle and bottom panels were swapped, as they were not matching the legend below.

[Read me is up to date](https://github.com/Mathilde-Reynes/ReynesAussel2024) with all the information necessary to run the model with these new changes.
Best,
