### 2025-03-01 #87 @heplesser
I commend Reynes and Aussel for their throrough and careful effort to reproduce the work of Bazhenov et al. Overall, this work already seems in quite good shape, but I think manuscript and code would benefit from some improvements.

First of all, I found it somewhat difficult to match figures in the manuscript to figures in the original paper. I would strongly suggest to mention in each figure caption the corresponding figure number in Bazhenov et al. I also wonder if it would not be easier to follow the reproduction step by step if figures were in the same order as in the original paper. I noticed that the manuscript contains excerpts of the figures in the original publication. While this definitely eases comparison, I wonder if this could raise licensing issues.

Concerning the results, I am most concerned about Fig 5 in the present manuscript. The reproduction results show approximately 20 up and down states compared to only 11 in the original paper, i.e., twice the switching frequency than in the original, not just "minimally shorter up-states" as commented on p. 14. Also, in the lower part of Fig 5, the initiation of the transition looks much sharper in the original figure than in the reproduction, with very straight opening "flanks". Furthermore, the original figure only shows broad stripes spanning the entire PY population, while the reproduction figure also includes "tongues" that include only parts of the PY population. I believe this merits further investigation. In that context, I noticed that Fig 13 in the manuscript shows a similar difference in patterns, although here related to the presence and absence of thalamocortical connections. Perhaps something similar could be underlying the differences in Fig 5?

Concerning the sustained activity initially shown in Fig 6 of the manuscript, it would be interesting to see the corresponding plot for the current implementation. I wonder if initial conditions could play a role here?

I would also like to point to a number of more detailed points.

1. You frequently refer to the ModelDB version of the Bazhenov et al model, but do not provide a reference for it. Please add the ModelDB accession number and a link.
2. I_{D->S} = -I_{S->D}, isn't it? It would be useful to point this out.
3. I think it would help the reader if you explicitly gave I_S^{int} somewhere above Eq 3 in the same way in which you give I_D^{int}.
4. Concerning Eqs 5, 6: Can the conductances g_D->S and g_S->D actually differ? Isn't this just the passive conductance connecting the dendritic and axosomatic compartments which must be symmetric for physical reasons?
5. On p 6, you give an expression for the mini rate µ including a parameter F, which I think is not given anywhere, and dividing by 250 at the very end of the expression. The corresponding expression on p 8694 in the original paper has F=400 and divides by 100, not 250. In the caption to Fig 11 you then have exactly the same numbers as in the original paper.
6. Does the connectivity include periodic boundary conditions or do neurons at the edge of the layer receive input only from the "inside" of the layer? Could you clarify this?
7. In Sec 4.2, it would benefit the reader if the text came before Table 2, so that one had an introduction to this large (and highly commendable!) table. I am still a bit confused about what you mean by "original value" and how you then obtained the often complex expression for the current model. Could you explain this more carefully? In the very first expression for the "present model", there seems to be a stray "B" at the end of the equation.
8. In Figs 2–4, I find the use of light blue vs dark blue problematic, since it provides only low contrast. Furthermore, differences are only visible if the light blue line extends beyond the dark blue—in principle difference in K could be "hiding" in Fig 2A.
9. I found your `environment.yml` file overspecified. As it includes explicit build hashes, it will at best work for the operating system on which it was generated. It is in a way useful to precisely describe the software you used for your simulations, but for users trying to use your model, you should provide a minimal environment file. I could run your code after creating a new environment with just `mamba create -n ra24 python numpy scipy matplotlib brian2`.
10. Your code has some very long lines (over 200 characters), which makes reading difficult even with today's wider screens. I would strongly suggest to break lines around 120 characters at the most (the old 80 limit seems dated). You may also want to run a linter or code formatter over the code.
11. I am confused by the code you use to introduce parameter variability, e.g. https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Cortical_layer.py#L117-L124. If I interpret the code correctly, you here aim to ensure that at most two parameters in a row receive a random adjustment with the same sign. This seems curious to me. Why not simply add random noise to the parameters as provided by the random number generator? Was something similar done in the original model? I did not see a comment on this in the manuscript.
12. On this line: https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Figure_plotting.py#L215 and possibly others, the string must be a raw string so that the \ will be passed correctly to LaTeX.
13. I find the connectivity specifications of the form https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Thalamo_cortical.py#L136 difficult to comprehend. At least at some point in the code there should be an explanatory comment on how these strings work.
14. I could successfully reproduce Fig 5, but when I set `fig_number = "1"` or 2, the simulation ran, but in the end reported only a mean firing rate of PY neurons of 17.678 Hz. What may have gone wrong here?

So much for now. Again, I am very impressed by the thorough work you have done in digging out details and parameters from old code and various papers and built a new implementation from this!
