### 2025-04-01 #87 @apdavison
The authors have done an excellent job in replicating this important study. As is usually the case in computational neuroscience (and sadly I don't think this situation has improved in the last 20 years), the published paper does not contain all the information needed to replicate the models and simulations described, so it is to the credit of Bazhenov and colleagues that they shared code on ModelDB, even if it does not correspond exactly to the precise version used in the original study.

Even with the availability of the ModelDB version, Reynes and Aussel's detective work was a non-trivial task, and the results - understanding and explaining the differences uncovered - are in general presented very clearly. I particularly appreciated Table 2, and the way the Results section was structured by the claims/assertions from the original paper.

## Reproducing the results

I would like to thank the authors for the modifications they made to the code, to make it easier to re-run simulations to match the figures.

I was able to run simulations for all figures, although the RNG seed available in `Thalamo_cortical.py` does not seem to be the one used for the manuscript. Figure 8, in particular, looks very different, presumably because of the highly stochastic and non-linear nature of up-state onset.

Since I don't use conda, I installed the project dependencies into a plain Python virtual environment (venv). For reference, the dependencies used were as follows (with Python 3.8.0):

```
Brian2==2.5.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
fonttools==4.55.3
importlib_resources==6.4.5
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
zipp==3.20.2
```

## Suggested modifications

1. I disagree that the full network dynamics in the replication are "very similar" to those in the original paper (p14). In particular, there are large differences in the frequency of the up-states. I suggest using "qualitatively similar" or just "similar".
2. Similarly, the up-states are not "minimally shorter", and the number of oscillations is not "negligibly" larger. On the basis of Figure 5, up-states in the replication seem to be smaller than in the original by a factor of 2 or more.
3. When citing using numbers, please ensure the sentence remains readable. For example "recreating [1] behavior" should be something like "recreating the behavior shown in the original paper [1]", and "using [1] C++ code" should be "using the original C++ code [1]".
4. The typesetting of equations and units was not standards-compliant, which hampered the readability of the document for me. Please modify your LaTeX source; you may find [this note](https://andrewdavison.info/notes/typesetting-comp-neuro/) helpful.
5. Table 2 should be introduced in the text before the table itself actually appears in the document.
6. I think the colormaps of membrane potentials would be better as grayscale. This would make comparison with the original figures easier.
7. There were many minor issues, such as typos. Rather than enumerate them all here, I have put my annotated copy of the manuscript [online here](https://sdrive.cnrs.fr/s/z4GyR6aFQwjKsRc) - please let me know if you are unable to access it.
