### 2025-04-04 #87 @Mathilde-Reynes
Response to @heplesser

We would like to thank the reviewer for their thorough and constructive review, and sincerely appreciate the time invested in assessing our manuscript. Their feedback raised valid points that helped us improve the clarity and robustness of both the manuscript and the code.
Below we address their comments and have indicated the associated changes in our point-by-point response. We also pushed the corrected manuscript on the git; with one version with changes in red and one ‘clean’ version.
______________________________________________________

> First of all, I found it somewhat difficult to match figures in the manuscript to figures in the original paper. I would strongly suggest to mention in each figure caption the corresponding figure number in Bazhenov et al. I also wonder if it would not be easier to follow the reproduction step by step if figures were in the same order as in the original paper. I noticed that the manuscript contains excerpts of the figures in the original publication. While this definitely eases comparison, I wonder if this could raise licensing issues.

•	We recognize that navigating the two papers simultaneously might be indeed laborious. We followed your advice of adding, in each figure caption, the corresponding figure number from Bazhenov et al. (2002).
•	We acknowledge your comment regarding the proposed rearrangement of the figures. However, we believe the order logically differs as the objectives of the two manuscripts are quite distinct. Bazhenov et al. (2002) aimed to present results and their interpretations in a narrative-driven manner. In our case, we focused on reproducing their claims with a more technical approach. Therefore, we structured our manuscript so that related claims are addressed within the same subsection, rather than switching back and forth as it would have been the case should we have followed Bazhenov et al. (2002) narrative. In that sense, we believe that adhering to the original paper's order would make the manuscript less readable and harder to follow.
•	As regards to potential licensing issues, as per our understanding of the website of original paper’s journal’s policy (https://www.jneurosci.org/content/rights-permissions), we believe we have permission to reuse its material, under the condition that we have to specify a full journal reference, link to the version of record, as well as a copyright statement (screen capture is provided below). We added all these elements as to respect JNeurosci guidelines to reuse their material.
**Modifications implemented on page 1.**

> Concerning the results, I am most concerned about Fig 5 in the present manuscript. The reproduction results show approximately 20 up and down states compared to only 11 in the original paper, i.e., twice the switching frequency than in the original, not just "minimally shorter up-states" as commented on p. 14. Also, in the lower part of Fig 5, the initiation of the transition looks much sharper in the original figure than in the reproduction, with very straight opening "flanks". Furthermore, the original figure only shows broad stripes spanning the entire PY population, while the reproduction figure also includes "tongues" that include only parts of the PY population. I believe this merits further investigation. In that context, I noticed that Fig 13 in the manuscript shows a similar difference in patterns, although here related to the presence and absence of thalamocortical connections. Perhaps something similar could be underlying the differences in Fig 5?

•	We agree that our statement “our model’s up‐states appear minimally shorter, resulting in negligibly more oscillations during a typical 30‐second simulation” underestimated the reality and we appreciate you pinpointing it. The manuscript was modified in that sense.
Modifications implemented on page 15.
•	Regarding the ‘sharper’ up-state initiation, we agree this is a significant difference between the present and the original model that we had not explicitly mentioned in the manuscript, and that we have corrected now.
Modifications implemented on page 15.
•	We have been further investigating these two points, as there were the two key points we struggled to fully comprehend. You are also correct in noting that Fig 13 looks more similar to the original; we had the same thought. That being said, since the original figure explicitly states the number of TC and RE cells and displays their activity—activity that does not exist when disconnected from the cortical layer—we believe this suggests that thalamocortical connections were indeed present in the original figure.
Reducing the weight of the specific connections from the thalamus to the cortical layer indeed slower oscillations, but not sufficiently enough to reproduce exactly the original results. For instance, running a simulation with only 10% of the original synaptic conductance between the thalamus and the cortical layer (g_syn_ampa_tcpy = 0.00001 msiemens; g_syn_ampa_tcin = 0.00001 msiemens) lead to 0.86 Hz cortical oscillations. This is still quite far from the 0.5Hz that can be observe in some figures of the original model.
Another potential approach, clearly illustrated in Figure 10 of our model, is to reduce the conductance of miniPSPs. However, in a fully connected model, while decreasing the amplitude of miniPSPs slightly lowers the frequency of up-states, it also disrupts their organization. For this reason, we do not consider this parameter a viable option for modification.
We also tested an increase in AMPA PY-IN conductance, as shown in Figure 12 of our manuscript, but this adjustment did not yield results closer to the original implementation.
Since the slow accumulation of the I_KCa current in the dendritic compartment is believed to facilitate the transition from an up-state to a down-state, we attempted to reduce it in order to prolong up-states and thereby slow cortical oscillations. By decreasing the conductance of I_KCa (g_kca = 0.3msiemenscm**-2 * 0.2), we observed a clear increase in the duration of cortical up-states, making them more similar to those in the original model. However, the oscillations remained too fast, with shorter down-states overall.
Nonetheless, these parameters cannot explain the discrepancy between our reproduction and the original model, as we use the same parameters when running the original C++ model and have verified that the calcium dynamics are consistent. As of now, we do not have an explanation for this difference between the original model and our implementation, still we acknowledge that this significant difference is important. We have added a supplementary figure showcasing the different tests we have performed to investigate that question.
**Modifications implemented on page 32**

> Concerning the sustained activity initially shown in Fig 6 of the manuscript, it would be interesting to see the corresponding plot for the current implementation. I wonder if initial conditions could play a role here?

•	When running the original model on our system, we were able to reproduce the initial sustained activity. However, we were unable to replicate this result in our own model. We suspect this discrepancy may be due to one of the following reasons: (1) differences in the compilers used (the original model is compiled with gcc from C++ code, while our model uses Cython), (2) variations in the initialization of our parameters, or (3) a potential error in the reproduction process, likely stemming from a parameter mismatch (that could also be responsible for the faster oscillations observed in our reproduction). As we believe that the absence of this initial surge in activity aligns more closely with the biological reality, we are comfortable with keeping the model as it is on that point. We have completed Figure 6 with two new graphs showing the initial activity of the various implementations of the model.
**Modifications implemented on page 16**

> I would also like to point to a number of more detailed points.
> 1.	You frequently refer to the ModelDB version of the Bazhenov et al model, but do not provide a reference for it. Please add the ModelDB accession number and a link.

•	Thank you for this remark. We have added the reference for the ModelDB, including an accession number and a link as per requested.

> 2.	I_{D->S} = -I_{S->D}, isn't it? It would be useful to point this out.

•	As we will develop in your point 4. below, as the conductances g_D->S and g_S->D differ, I_{D->S} does not equal -I_{S->D}. Please refer to 4. for an explanation and the relevant changes made in the manuscript.

> 3.	 I think it would help the reader if you explicitly gave I_S^{int} somewhere above Eq 3 in the same way in which you give I_D^{int}.

•	I_S^{int} is described in Eq 4 just below Eq 3.

> 4.	Concerning Eqs 5, 6: Can the conductances g_D->S and g_S->D actually differ? Isn't this just the passive conductance connecting the dendritic and axosomatic compartments which must be symmetric for physical reasons?

•	In the present model g_D->S and g_S->D actually differ.
Actually, g is defined as 1/(R) where R = r*s_dend for the junction between the soma and the dendrite and R=r*s_soma for the junction between the dendrites and the soma. r here is the resistance between compartments (10MOhm) and s_dend and s_soma are the area of the specific compartments.  Therefore, the difference in compartment sizes is the main reason why g_D->S and g_S->D differ. Indeed, for PY neurons, the dendrite area is 165 times larger than the soma, and for IN neurons, it is 50 times bigger.  If the same conductance density is used for both compartments, the dendritic compartment has a much higher absolute conductance due to its larger area, and therefore a lower input resistance. Basically this asymmetry translates the fact that somas are more reactive than dendrites. Indeed here, for the same injected current, the voltage change in the soma will be significatively higher than in the dendrite. This makes the soma more sensitive to dendritic inputs. In order to clarify this, we have added in the manuscript the equations for g_D->S and g_S->D.
**Modifications implemented page 5**

> 5.	On p 6, you give an expression for the mini rate µ including a parameter F, which I think is not given anywhere, and dividing by 250 at the very end of the expression. The corresponding expression on p 8694 in the original paper has F=400 and divides by 100, not 250. In the caption to Fig 11 you then have exactly the same numbers as in the original paper.

•	You're right to point out that equation. Not only is the "250" a typo, but the equation presented here uses an exponential arrival rate, whereas the rest of the paper consistently implements a logarithmic arrival rate. We have corrected this in the manuscript, ensuring the correct equation is shown and explicitly specifying the value of F.
Modifications implemented page 6

> 6.	Does the connectivity include periodic boundary conditions or do neurons at the edge of the layer receive input only from the "inside" of the layer? Could you clarify this?

•	In our current implementation, we have not set a specific connectivity boundary condition. Therefore, neurons at the edges of the layer receive input from fewer neurons compared to the neurons located further away from the edges. If we are not mistaken, this aligns with the second part of your question, stating that edge neurons “receive input only from the inside of the layer.”
The original paper does not discuss boundary conditions. However, in the ModelDB implementation, a function is provided to define the desired boundary condition, allowing for options such as flow boundary condition, periodic boundary condition or no boundary condition (parameter BOUND and BOUNDcx, in function b()). The ModelDB version was configured with a parameter BOUND == 0 and BOUNDcx == 0, indicating that a flow boundary condition was used. However, since the manuscript does not explicitly describe it and as per our remarks on other parameters in this paper, we cannot positively confirm whether this was the condition applied in the final figures of the manuscript.
Given that we did not observe a significant impact of the boundary condition on our results, we chose not to impose a periodic boundary condition. We have added detail on boundary condition in the table of the manuscript.
**Modifications implemented page 7 & 10**

> 7.	In Sec 4.2, it would benefit the reader if the text came before Table 2, so that one had an introduction to this large (and highly commendable!) table. I am still a bit confused about what you mean by "original value" and how you then obtained the often complex expression for the current model. Could you explain this more carefully? In the very first expression for the "present model", there seems to be a stray "B" at the end of the equation.

•	Thank you for the highly relevant remark. We put the text before the table in order to facilitate comprehension as you proposed. We added an extra sentence in this text to explain more thoroughly what we develop in the table below it. We removed the stray ‘B’ which was indeed a typo.
**Modifications implemented page 8**

> 8.	In Figs 2–4, I find the use of light blue vs dark blue problematic, since it provides only low contrast. Furthermore, differences are only visible if the light blue line extends beyond the dark blue—in principle difference in K could be "hiding" in Fig 2A.

•	We replotted the figure 2-4 with different colors in order to maximize the contrast. However, we would like to point out that whatever color is used to plot, if the plots align perfectly, one would only see one of the two colors. We played with different line thickness and transparency but they only slightly facilitate the readability of the figure. In order to explicitly confirm that there exist no different in K, we added a zoomed-in plot on the right of figure, that we think combined with our sentence “Should the equations and their integration be the same, a perfect match between the two will be observed. This is the case for $I_K$ but not $I_{Na}$.”; should make our statement clearer.
**Modifications implemented page 12**

> 9.	I found your environment.yml file overspecified. As it includes explicit build hashes, it will at best work for the operating system on which it was generated. It is in a way useful to precisely describe the software you used for your simulations, but for users trying to use your model, you should provide a minimal environment file. I could run your code after creating a new environment with just mamba create -n ra24 python numpy scipy matplotlib brian2.

•	Thank you for your feedback, indeed the environment file should be minimal. We have provided a new environment file that follows this guideline.
**Modifications implemented on the git**

> 10.	Your code has some very long lines (over 200 characters), which makes reading difficult even with today's wider screens. I would strongly suggest to break lines around 120 characters at the most (the old 80 limit seems dated). You may also want to run a linter or code formatter over the code.

**•	The code is now formatted using the formatter Black.**

> 11.	I am confused by the code you use to introduce parameter variability, e.g. https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Cortical_layer.py#L117-L124. If I interpret the code correctly, you here aim to ensure that at most two parameters in a row receive a random adjustment with the same sign. This seems curious to me. Why not simply add random noise to the parameters as provided by the random number generator? Was something similar done in the original model? I did not see a comment on this in the manuscript.

•	You are correct we did not explain the parameter variability. We explicitly added a paragraph on this in the reviewed manuscript. This random adjustment is strictly reproducing the one in the original model.
**Modifications implemented page 6**

> 12.	On this line: https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Figure_plotting.py#L215 and possibly others, the string must be a raw string so that the \ will be passed correctly to LaTeX.

•	This is now corrected.

> 13.	I find the connectivity specifications of the form https://github.com/Mathilde-Reynes/ReynesAussel2024/blob/8de8bb5df75d3f1eb4f7fcd8d7c5ed014e4beacc/Model/Thalamo_cortical.py#L136 difficult to comprehend. At least at some point in the code there should be an explanatory comment on how these strings work.

•	As proposed, we have added explanatory comments for the connectivity specifications on all files with synapses (Cortical_layer.py ; Thalamus.py and Thalamo_cortical.py).

> 14.	I could successfully reproduce Fig 5, but when I set fig_number = "1" or 2, the simulation ran, but in the end reported only a mean firing rate of PY neurons of 17.678 Hz. What may have gone wrong here?

•	Basically, as of this submission, whenever a user provided the function with a number that is not an actual figure that can be plotted (such as 1 or 2), a 10-seconds simulation with the base parameters (the ones of Figure 5) is ran. An additional warning message was added in the code to make the situation clearer.
•	The mean firing rate of pyramidal neurons is computed by dividing the total number of recorded spikes (len(R2_PYs.t)) by the number of PY neurons (N_PY) and the total simulation runtime (runtime). The mean firing rate of PY neurons is different from the slow oscillations frequency. Indeed, during an upstate, PY neurons spike multiple times (could be 15 to 25 times), thus leading to the 17Hz mean firing rate that you reported. According to us, nothing went wrong in the simulation you ran.

______________________________________________________

We would like to thank the reviewer again for the constructive comments and remarks, and are glad you can appreciate the effort we put in reproducing this model.
