### 2025-04-04 #87 @Mathilde-Reynes
Response to @apdavison

We sincerely thank the reviewer for their initial review on January 10th, which prompted us to refine our code, making it more user-friendly and facilitating easier reproduction of results. We would like also to thank the reviewer for their latest review, and sincerely appreciate the time invested in assessing our manuscript. Their feedback raised valid points that helped us improve the clarity, readability and robustness of both the manuscript and the code.
Below we address their comments and have indicated the associated changes in our point-by-point response. We also pushed the corrected manuscript on the git; with one version with changes in red and one ‘clean’ version.
______________________________________________________

> […] I was able to run simulations for all figures, although the RNG seed available in Thalamo_cortical.py does not seem to be the one used for the manuscript. Figure 8, in particular, looks very different, presumably because of the highly stochastic and non-linear nature of up-state onset.

•	Indeed, the seed used has a significant impact on the spatiotemporal dynamics of cortical oscillations, as the miniature excitatory postsynaptic potentials (modeled as a Poisson process) influence the onset of up-states in the oscillations. Additionally, variability in some conductances may also contribute to differences, and we have added a paragraph addressing this variability in the revised manuscript as we did not mention it in the initial submission and as reviewer 1 suggested it (see modifications on page 6). However, regardless of the seed, the qualitative behavior of the model remains consistent.
The RNG seed available in Thalamo_cortical.py was in fact not the one used for this specific figure of the manuscript and we apologize for this. Running the model with the conda env on the git, with the RNG seed currently available, we were not able to reproduce the figure you sent (panel A of the figure below). This same result was obtained on a Linux and a Windows machine. In order to investigate, we have installed your venv on my Windows machine and ran the model on this environment. I still obtain the same results (panel B of the figure below). Thus we could not reproduce the issue with the elements mentioned.
We would like to emphasize that some compiler optimizations methods can vary floating-point precision and may lead to such differences. For example, should you allow the compiler to reorder your arithmetic operations and optimize to get faster math instructions, results can vary in complex non-linear models.
Therefore, even if we ensure the seed is fixed, we cannot guarantee exact reproducibility of results—especially in the longer term or across different environments.
**Modifications implemented page 19**

![Image](https://github.com/user-attachments/assets/ec98df1d-6c0d-4b6d-b19a-4ad4842526ed)

> Since I don't use conda, I installed the project dependencies into a plain Python virtual environment (venv). For reference, the dependencies used were as follows (with Python 3.8.0) […].

•	Thank you for providing us with your verv specifications. We have added, in the read me of the git, a note for users that would like to run the model without conda. This should help reproducibility as well.
**Modifications implemented on the git**

> Suggested modifications:
> 1.	I disagree that the full network dynamics in the replication are "very similar" to those in the original paper (p14). In particular, there are large differences in the frequency of the up-states. I suggest using "qualitatively similar" or just "similar".
> 2.	Similarly, the up-states are not "minimally shorter", and the number of oscillations is not "negligibly" larger. On the basis of Figure 5, up-states in the replication seem to be smaller than in the original by a factor of 2 or more.

•	Thank you for these two very relevant remarks (1 and 2). We agree that our statements “our model’s up‐states appear minimally shorter, resulting in negligibly more oscillations during a typical 30‐second simulation” and “the reproduced model displays dynamics very similar to those described by Bazhenov et al.” totally underestimated the reality of our results and we appreciate you pinpointing it. The manuscript was modified in that sense.
**Modifications implemented page 15**

> 3.	When citing using numbers, please ensure the sentence remains readable. For example "recreating [1] behavior" should be something like "recreating the behavior shown in the original paper [1]", and "using [1] C++ code" should be "using the original C++ code [1]".

•	Changes have been made to the manuscript to improve its fluidity and readability. Given the extensive modifications across multiple pages, we invite you to consult the ‘reviewed’ version of the manuscript, where all changes are highlighted in red. Basically, we have incorporated additional context whenever references are mentioned in the text, as per your recommendation. These changes especially apply for citations of the original paper and ModelDB.

> 4.	The typesetting of equations and units was not standards-compliant, which hampered the readability of the document for me. Please modify your LaTeX source; you may find [this note](https://andrewdavison.info/notes/typesetting-comp-neuro/) helpful.

•	Thank you for providing the link to this, extremely useful, note. We applied your recommendations to all our equations and units so that they are standards-compliant. We did not highlighted them in red as basically all the equations and units were changed.

> 5.	Table 2 should be introduced in the text before the table itself actually appears in the document.

•	Thank you for the highly relevant remark. We put the text before the table in order to facilitate comprehension as you proposed. We added an extra sentence in this text to explain more thoroughly what we develop in the table below it.
**Modifications implemented page 8**

> 6.	I think the colormaps of membrane potentials would be better as grayscale. This would make comparison with the original figures easier.

•	The colormaps were replotted as grayscale.
**Modifications implemented page 22 & 15**

> 7.	There were many minor issues, such as typos. Rather than enumerate them all here, I have put my annotated copy of the manuscript [online here](https://sdrive.cnrs.fr/s/z4GyR6aFQwjKsRc) - please let me know if you are unable to access it.

•	We were able to access the annotated copy of the manuscript and made all the changes you noted. We thank you for providing feedback on this aspect of our work as well, which is definitely as important as the scientific content itself.

______________________________________________________

We would like to thank the reviewer again for the very helpful comments and remarks, and are glad you can appreciate the effort we put in reproducing this model.
