### 2025-04-05 #87 @rougier
@Mathilde-Reynes Thanks for the detailed answer.
@heplesser @apdavison Are you satisfied with the answers and corrections?
