### 2025-04-28 #87 @apdavison
I am satisfied with the answers and corrections.

I do have one further query for the authors: now that the colormaps are in grayscale, it seems as though there is a systematic difference between the original and replicated versions (e.g., Figure 5): specifically, both up- and down-states appear substantially darker in the replications. Does this represent an actual difference in the activity levels, or is it just that the exact gray-level map is not known for the original figures?
