### 2025-04-28 #87 @Mathilde-Reynes
@apdavison,
Thank you for the very relevant remark. To better understand the situation, let’s refer to Figure 7 of our manuscript:

![Image](https://github.com/user-attachments/assets/cf1f69c2-7ed0-4604-8151-9c92821d1fe9)

In the bottom panel of Figure 7, we see that the resting membrane potential appears to be consistent across all cell types between the original model and our implementation, even though the lines in the original paper’s plot are missing (-68 mV for PY cells, -66 mV for IN cells, -73 mV for RE, and -72 mV for TC). According to this, the voltage range during the down state is similar in both our implementation and the original model. That being said, the colors in Figure 5 during the down states should therefore match, but they do not. This suggests that either the colormap or the cbar used in both plots are not identical, as you suggested. We did attempt to find a more suitable grayscale to align with the original figures, but without success.

Your question raised an important point: why does the down state for pyramidal cells appear lighter (whiter) than that of thalamic cells in the original manuscript, even though its value is higher (-68 mV vs. -72 mV)? The down-state value of -72 mV for TC cells should logically appear lighter than the -68 mV value for PY cells, yet this is not the case. This likely indicates that _two different colorbars were used across the same figure_, which would also explain why we couldn’t find an appropriate color scale to replicate it.
To illustrate this, we’ve re-plotted Figure 5 with two distinct colorbars for pyramidal and thalamic relay cells, using arbitrary maximum and minimum values for the colorbars. This version aligns much better with the original manuscript’s figure though it is less intuitive for readers who can no longer easily compare the two cell types. Therefore, we do not believe using two separate color scales is the best approach.

![Image](https://github.com/user-attachments/assets/3f13d3fc-6da0-40fd-9b2a-da134ef3cefd)

Since we think this explanation will be helpful to the reader, we have added a brief sentence in the manuscript to summarize the point discussed above:
“Nonetheless, our figure still differs because the original figure likely used two different colorbars for pyramidal and thalamic cells, resulting in internal mismatches that naturally explain the differences with our figure which uses a consistent colorbar across all plots. This claims is supported by the cellular activity shown in Figure 7.”
**Modifications implemented page 15**
