### 2025-04-30 #87 @rougier
@apdavison Thansk, do that means the article is now ok for you?
@heplesser Are you satisfied with answers and corrections?
