### 2025-04-30 #87 @apdavison
@rougier yes: for me the article is good to publish
