### 2025-05-07 #87 @rougier
@heplesser Are you ok with answers and corrections?
