### 2025-05-15 #87 @heplesser
@Mathilde-Reynes Thank you very much for the thorough revisions to your manuscript and your careful and detailed answers to my questions and the code improvements. I have a few more comments, mostly minor, except possibly the last:

1. There seems to be no Figure 1 in your manuscript. The first figure I could find is Figure 2 on p. 12. Is this possibly and artifact of the ReScience LaTeX style, or has a figure gone missing?
2. The programming language should be capitalized: Python
3. p 3: I am not sure "sustainment" is a word. I think "Sustained up-state activity" would be better.
4. The first block of equations on p 5 contains six lines of equations but only four equation numbers. This seems confusing. Also, use \times as multiplication symbol, not the asterisk.
5. p 6: I would suggest "Parameter variability" or maybe even more specific here "Parameter randomization"
6. On p 15, you write "Following thorough investigation, we have yet to properly elucidate this inconsistency in our model. Figure S4 presents our attempts to adjust parameters to better align with the behavior of the original model." I think it would be even better if you described very briefly here (a) that running the code from ModelDB resulted in dynamics closer to the original figure (this is at least my impression from Fig S4A), which parameter modifications you attempted and what that achieved, and then pointed to Fig S4 for details. Now, the important information is hidden in the figure caption in the appendix. Given that the ModelDB code seems to produce dynamics closer to the original, you may also want to suggest that one might do further investigations on the original code and yours to uncover the source of the differences. But just mention it, I do not mean to suggest that you investigate this now.
7. You reference S4 on p 15, before at least S3, which is first referenced on p 25. Please order the figures in the appendix in order of first reference.
8. In my first review, I seem to have overlooked a major difference in time scales in Figs 9 and 11. In both figures, your traces are 8s long, while the scale bars in the original figures suggest that the traces there are about 100s and 40s long, respectively. Just imagining your figures extended to 100s and 40s, respectively, I find it difficult to see that they would agree particularly well with the original figures. Could you comment on this? I tried to create Fig 9 for a longer time stretch, but while I managed to increase the runtime, the plot still only showed the first 8 seconds.

Best regards!
