### 2025-06-02 #87 @Mathilde-Reynes
@heplesser
Once again thank you for the comments and remarks. Below we address the comments and have indicated the associated changes in our point-by-point response. We also pushed the corrected manuscript on the git; with one version with changes in red and one ‘clean’ version.
______________________________________________________

> 1. There seems to be no Figure 1 in your manuscript. The first figure I could find is Figure 2 on p. 12. Is this possibly and artifact of the ReScience LaTeX style, or has a figure gone missing?

Thank you for pointing this out. We confirm that Figure 1 is present in the latest submission and is titled “Example of connectivity between populations varying in size.” However, we acknowledge that its appearance—closer to a table— make it less immediately recognizable as a figure.

> 2. The programming language should be capitalized: Python

Thank you for this comment, the manuscript was modified accordingly.
**Modifications implemented on pages 16 & 31.**

> 3. p 3: I am not sure "sustainment" is a word. I think "Sustained up-state activity" would be better.

Thank you for the remark. While "sustainment" does exist ([as referenced in the Oxford English Dictionary](https://www.oed.com/dictionary/sustainment_n?tab=factsheet#19474260), particularly in US English), it is indeed rarely used. For the sake of clarity and readability, we have replaced it with "sustained up-state activity" as suggested.
**Modifications implemented on page 3.**

> 4. The first block of equations on p 5 contains six lines of equations but only four equation numbers. This seems confusing. Also, use \times as multiplication symbol, not the asterisk.

Thank you for this comment, the manuscript was modified accordingly.
**Modifications implemented on page 5.**

> 5. p 6: I would suggest "Parameter variability" or maybe even more specific here "Parameter randomization"

We have changed the title to “Parameter randomization”.
**Modifications implemented on page 6.**

> 6. On p 15, you write "Following thorough investigation, we have yet to properly elucidate this inconsistency in our model. Figure S4 presents our attempts to adjust parameters to better align with the behavior of the original model." I think it would be even better if you described very briefly here (a) that running the code from ModelDB resulted in dynamics closer to the original figure (this is at least my impression from Fig S4A), which parameter modifications you attempted and what that achieved, and then pointed to Fig S4 for details. Now, the important information is hidden in the figure caption in the appendix. Given that the ModelDB code seems to produce dynamics closer to the original, you may also want to suggest that one might do further investigations on the original code and yours to uncover the source of the differences. But just mention it, I do not mean to suggest that you investigate this now.

Thank you for the suggestion. We have revised the text on page 15 to briefly summarize that running the original ModelDB code produced dynamics more consistent with those shown in the original figure. We now also mention the key parameter modifications we attempted, before directing the reader to Figure S4 for details. Finally, we added a sentence suggesting that further investigations comparing the original and reproduced implementations may help clarify the source of the remaining discrepancies.
**Modifications implemented on pages 15 & 16.**

> 7. You reference S4 on p 15, before at least S3, which is first referenced on p 25. Please order the figures in the appendix in order of first reference.

We have reordered the supplementary figures so that they now appear in the order of their first reference in the main text.

> 8. In my first review, I seem to have overlooked a major difference in time scales in Figs 9 and 11. In both figures, your traces are 8s long, while the scale bars in the original figures suggest that the traces there are about 100s and 40s long, respectively. Just imagining your figures extended to 100s and 40s, respectively, I find it difficult to see that they would agree particularly well with the original figures. Could you comment on this? I tried to create Fig 9 for a longer time stretch, but while I managed to increase the runtime, the plot still only showed the first 8 seconds.

We have also noted this apparent discrepancy. Indeed, if the scale bars in Figures 9 and 11 of our manuscript were truly 25 seconds and 10 seconds long, respectively, we would agree that the dynamics shown in our implementation would significantly deviate from those in the original study. However, upon close inspection of the original figures, we believe there may be a mislabeling of the time scales.
For instance, in the original manuscript's Figure 4 (which corresponds to our Figure 9), the stated scale bar of 25 seconds appears inconsistent with the rest of the paper’s content. If that scale were accurate, the observed up-states in the N = 100 cell condition (used throughout most of the original paper) would occur roughly 8 times in 100 seconds—corresponding to a frequency of only 0.08 Hz. This is far lower than the typical range of cortical slow oscillations, and very far from what is shown in other figures of the original manuscript as well. A more plausible interpretation is that the scale bar actually represents 2.5 seconds, not 25, which would imply a frequency of approximately 0.8 Hz—much more consistent with cortical dynamics discussed in the manuscript. This scale of 2.5 seconds would actually match the one used in Figure 7 of the original manuscript.
A similar interpretation applies to our Figure 11. If the full trace is 10 seconds long (rather than 40), it would represent oscillations at ~1 Hz, again aligning better with expected physiological behavior.
We have added notes on the two figures concerned in order to quickly summarize the assumptions made here.
Regarding your second point, it is normal that increasing the runtime alone does not affect the duration shown in Figure 9. This is because the figure was specifically designed to reproduce the exact output displayed in the original manuscript. For those interested in generating custom plots or extending the simulation duration, we recommend directly accessing the relevant results from Thalamo_cortical.py, which allows more flexible control over the plotting.
**Modifications implemented on pages 20 & 21.**

______________________________________________________
Best regards
