### 2025-06-03 #87 @rougier
@Mathilde-Reynes Thank you for your details answer and corrections.
@heplesser Are you satisfied with answers and corrections ?
