### 2025-06-11 #87 @heplesser
@Mathilde-Reynes Thank you very much for your revisions and reply. I am still not persuaded concerning the scale bars on Bazhenov et al's Fig 4 and your Fig 9. Note that Bazhenov et al on p 8695, right column, write that a 20PY-4IN network "produc[ed] a low rate of bursting (25–50 sec intervals between bursts) and a high variability of interburst intervals". Their Fig 4A, left panel top, is consistent with this description provided the scale bar is 25 sec long, as shown in the figure. While their figure only shows a single interburst interval making it difficult to judge what they mean by "high variability", the interburst intervals in your Fig 9A top left seem rather regular to me. Bazhenov et al further write that a 100PY-25IN "network fired more regularly and at a hight frequency of ~0.05 Hz", which seems consistent with the 0.08 Hz you mention. Furthermore, the bursts in your Fig 9A left seem to last much longer, compared to the periods of silence, than in Bazhenov's data: Your burst are as long if not longer than the pauses, while Bazhenov's bursts in the left panel of their Fig 4A are almost all very narrow spikes with significantly wider silences in between. So I am afraid that something is more seriously off here. If you recreate these figures with the ModelDB code, what do you find and what timescale? Although, looking at the C code from ModelDB now, running this simple PY-IN only model seems not trivial.

@rougier I am wondering a bit about ReScience policy here: Are clear deviations from the original acceptable in a ReScience paper? If so, I would only require an adjusted discussion of Fig 9 before accepting. If reproduction of the original results is needed, the underlying problem would need to be resolved before acceptance.

Best regards
