### 2025-06-14 #87 @rougier
@heplesser Yes, we can accept deviations from the original as long as it is motivated. It is indeed not rare to have such deviations, worst case scenario being to flag the original study to be not replicable.  I would thus advise to adjust the discussion. If a reader has a doubt, he/she can have a look at the review (here) and possibly ask questions after publication.

In the meantime, @Mathilde-Reynes can you have a look at what @heplesser suggested?
