### 2025-06-30 #87 @Mathilde-Reynes
@heplesser
Thank you again for your thorough investigation of our manuscript.
__________________________
> So I am afraid that something is more seriously off here. If you recreate these figures with the ModelDB code, what do you find and what timescale? Although, looking at the C code from ModelDB now, running this simple PY-IN only model seems not trivial.

Following your advice, we have recreated Figure 4 from Bazhenov et al's article (our Figure 9) using the C code from ModelDB, and we can now confirm that the scale bars in the original manuscript were indeed wrong (and therefore, so was the text describing the interburst interval). We have added our replication of Bazhenov et al's figure as supplementary material (Figure S4) to support our claim.
While this is hard to check, we think that the figure in the original manuscript was obtained with a slightly different code, in which the conductances of synapses and miniEPSPs were not scaled with the number of neurons (as explained on page 20 of our article). The original manuscript also states: _"we increased the amplitude of miniature events by ∼50% in our simulations to obtain SWS-like oscillations in relatively small networks"_, so it is seems likely that Figure 4 was generated with a smaller mini amplitude compared to the code available on ModelDB, resulting in increased interburst interval. However, without the original values of mini amplitude used, we cannot exactly replicate this experiment.
**Modifications implemented on pages 20, 21 & 33.**
___________________________
Best regards
