### 2025-07-01 #87 @heplesser
@Mathilde-Reynes Thank you for your further investigation. I believe this provides enough information for anyone planning to pursue further work with this model. Thank you for your major efforts!

@rougier I now recommend the paper for acceptance and publication!
