### 2025-07-04 #87 @rougier
@heplesser @apdavison Thank you again for your very thorough review that helped improving the paper
@Mathilde-Reynes *Congratulations*, your submission is now officially accepted and I'll (try to) publish it soon even though this is not he best period for me. At this point, you can help by editing the `metadata.yaml` file with all the information you can find (for some others, you can't unless I provided the information).
