### 2025-07-04 #87 @rougier
I will need to access your sources for the article. If they're on overleaf, send me (by mail, not here or else anybody will be able to edit) the link. If they're on GitHub, you ca post the repo here.
