### 2025-07-22 #87 @rougier
Sorry for the delay, I got some problems with your latex files. The **sandbox** version is [here](https://sandbox.zenodo.org/record/294242) (not official yet). Can you check eveythign is ok? (DOI will be changed later)
