### 2025-07-25 #87 @Mathilde-Reynes
@rougier apologies for the delay.

We've just realized there's a small typo we made at the top of page 5 :
currently this is written: `$area_{\mathrm{dend}_{\mathrm{PY}}} = 50\times area_{\mathrm{soma}}$ for IN cells. `
**but it should be: `$area_{\mathrm{dend}_{\mathrm{IN}}} = 50\times area_{\mathrm{soma}}$ for IN cells. `**

Other than that, everything looks fine. Thank you, and we apologize for the small typo.
