### 2025-07-28 #87 @rougier
Arg, it will delay the publication by at least one month!
Just kidding, the article is onlien at https://zenodo.org/record/16529268 and will soon appear on the website.
Congratulations!
